test: split list selection test
